+++ prql-compiler/src/tests/mod.rs
@@ -1,5 +1,6 @@
 mod test;
 mod test_bad_error_messages;
+mod test_codegen;
 mod test_error_messages;
 
 pub use test::compile;
